+++ packages/bsky/src/api/app/bsky/actor/getProfiles.ts
@@ -18,7 +18,7 @@ export default function (server: Server, ctx: AppContext) {
     handler: async ({ auth, params, req }) => {
       const viewer = auth.credentials.iss
       const labelers = ctx.reqLabelers(req)
-      const hydrateCtx = { viewer, labelers }
+      const hydrateCtx = await ctx.hydrator.createContext({ viewer, labelers })
 
       const result = await getProfile({ ...params, hydrateCtx }, ctx)
 
@@ -29,7 +29,7 @@ export default function (server: Server, ctx: AppContext) {
         body: result,
         headers: resHeaders({
           repoRev,
-          labelers,
+          labelers: hydrateCtx.labelers,
         }),
       }
     },