+++ packages/bsky/src/api/app/bsky/feed/getListFeed.ts
@@ -26,7 +26,7 @@ export default function (server: Server, ctx: AppContext) {
     handler: async ({ params, auth, req }) => {
       const viewer = auth.credentials.iss
       const labelers = ctx.reqLabelers(req)
-      const hydrateCtx = { labelers, viewer }
+      const hydrateCtx = await ctx.hydrator.createContext({ labelers, viewer })
 
       const result = await getListFeed({ ...params, hydrateCtx }, ctx)
 
@@ -35,7 +35,7 @@ export default function (server: Server, ctx: AppContext) {
       return {
         encoding: 'application/json',
         body: result,
-        headers: resHeaders({ labelers, repoRev }),
+        headers: resHeaders({ labelers: hydrateCtx.labelers, repoRev }),
       }
     },
   })