+++ packages/bsky/tests/data-plane/subscription/util.test.ts
@@ -1,10 +1,10 @@
 import { wait } from '@atproto/common'
+import { randomStr } from '@atproto/crypto'
 import {
   ConsecutiveList,
   LatestQueue,
   PartitionedQueue,
 } from '../../../src/data-plane/server/subscription/util'
-import { randomStr } from '../../../../crypto/src'
 
 describe('subscription utils', () => {
   describe('ConsecutiveList', () => {