+++ packages/dev-env/src/mock/index.ts
@@ -1,10 +1,6 @@
 import { AtUri } from '@atproto/syntax'
-import AtpAgent from '@atproto/api'
+import AtpAgent, { COM_ATPROTO_MODERATION } from '@atproto/api'
 import { Database } from '@atproto/bsky'
-import {
-  REASONSPAM,
-  REASONOTHER,
-} from '@atproto/api/src/client/types/com/atproto/moderation/defs'
 import { EXAMPLE_LABELER, TestNetwork } from '../index'
 import { postTexts, replyTexts } from './data'
 import labeledImgB64 from './img/labeled-img-b64'
@@ -15,13 +11,12 @@ import blurHashB64 from './img/blur-hash-avatar-b64'
 // we use this to ensure the mock dataset is always the same
 // which is very useful when testing
 // (not everything is currently deterministic but it could be)
-function* dateGen() {
+function* dateGen(): Generator<string, never> {
   let start = 1657846031914
   while (true) {
     yield new Date(start).toISOString()
     start += 1e3
   }
-  return ''
 }
 
 export async function generateMockSetup(env: TestNetwork) {
@@ -119,7 +114,10 @@ export async function generateMockSetup(env: TestNetwork) {
   // Report one user
   const reporter = picka(users)
   await reporter.agent.api.com.atproto.moderation.createReport({
-    reasonType: picka([REASONSPAM, REASONOTHER]),
+    reasonType: picka([
+      COM_ATPROTO_MODERATION.DefsReasonSpam,
+      COM_ATPROTO_MODERATION.DefsReasonOther,
+    ]),
     reason: picka(["Didn't look right to me", undefined, undefined]),
     subject: {
       $type: 'com.atproto.admin.defs#repoRef',
@@ -169,7 +167,10 @@ export async function generateMockSetup(env: TestNetwork) {
     if (rand(6) === 0) {
       const reporter = picka(users)
       await reporter.agent.api.com.atproto.moderation.createReport({
-        reasonType: picka([REASONSPAM, REASONOTHER]),
+        reasonType: picka([
+          COM_ATPROTO_MODERATION.DefsReasonSpam,
+          COM_ATPROTO_MODERATION.DefsReasonOther,
+        ]),
         reason: picka(["Didn't look right to me", undefined, undefined]),
         subject: {
           $type: 'com.atproto.repo.strongRef',