+++ packages/dev-env/src/seed/client.ts
@@ -1,11 +1,12 @@
 import fs from 'fs/promises'
 import { CID } from 'multiformats/cid'
-import AtpAgent from '@atproto/api'
-import { Main as Facet } from '@atproto/api/src/client/types/app/bsky/richtext/facet'
-import { InputSchema as CreateReportInput } from '@atproto/api/src/client/types/com/atproto/moderation/createReport'
-import { Record as PostRecord } from '@atproto/api/src/client/types/app/bsky/feed/post'
-import { Record as LikeRecord } from '@atproto/api/src/client/types/app/bsky/feed/like'
-import { Record as FollowRecord } from '@atproto/api/src/client/types/app/bsky/graph/follow'
+import AtpAgent, {
+  ComAtprotoModerationCreateReport,
+  AppBskyFeedPost,
+  AppBskyRichtextFacet,
+  AppBskyFeedLike,
+  AppBskyGraphFollow,
+} from '@atproto/api'
 import { AtUri } from '@atproto/syntax'
 import { BlobRef } from '@atproto/lexicon'
 import { TestNetworkNoAppView } from '../network-no-appview'
@@ -191,7 +192,11 @@ export class SeedClient<
     return this.profiles[by]
   }
 
-  async follow(from: string, to: string, overrides?: Partial<FollowRecord>) {
+  async follow(
+    from: string,
+    to: string,
+    overrides?: Partial<AppBskyGraphFollow.Record>,
+  ) {
     const res = await this.agent.api.app.bsky.graph.follow.create(
       { repo: from },
       {
@@ -218,7 +223,11 @@ export class SeedClient<
     delete this.follows[from][to]
   }
 
-  async block(from: string, to: string, overrides?: Partial<FollowRecord>) {
+  async block(
+    from: string,
+    to: string,
+    overrides?: Partial<AppBskyGraphFollow.Record>,
+  ) {
     const res = await this.agent.api.app.bsky.graph.block.create(
       { repo: from },
       {
@@ -248,10 +257,10 @@ export class SeedClient<
   async post(
     by: string,
     text: string,
-    facets?: Facet[],
+    facets?: AppBskyRichtextFacet.Main[],
     images?: ImageRef[],
     quote?: RecordRef,
-    overrides?: Partial<PostRecord>,
+    overrides?: Partial<AppBskyFeedPost.Record>,
   ) {
     const imageEmbed = images && {
       $type: 'app.bsky.embed.images',
@@ -315,7 +324,11 @@ export class SeedClient<
     return { image: res.data.blob, alt: filePath }
   }
 
-  async like(by: string, subject: RecordRef, overrides?: Partial<LikeRecord>) {
+  async like(
+    by: string,
+    subject: RecordRef,
+    overrides?: Partial<AppBskyFeedLike.Record>,
+  ) {
     const res = await this.agent.api.app.bsky.feed.like.create(
       { repo: by },
       {
@@ -335,7 +348,7 @@ export class SeedClient<
     root: RecordRef,
     parent: RecordRef,
     text: string,
-    facets?: Facet[],
+    facets?: AppBskyRichtextFacet.Main[],
     images?: ImageRef[],
   ) {
     const embed = images
@@ -447,8 +460,8 @@ export class SeedClient<
   }
 
   async createReport(opts: {
-    reasonType: CreateReportInput['reasonType']
-    subject: CreateReportInput['subject']
+    reasonType: ComAtprotoModerationCreateReport.InputSchema['reasonType']
+    subject: ComAtprotoModerationCreateReport.InputSchema['subject']
     reason?: string
     reportedBy: string
   }) {