+++ packages/aws/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @atproto/aws
 
+## 0.2.0
+
+### Minor Changes
+
+- [#2169](https://github.com/bluesky-social/atproto/pull/2169) [`f689bd51a`](https://github.com/bluesky-social/atproto/commit/f689bd51a2f4e02d4eca40eb2568a1fcb95494e9) Thanks [@matthieusieben](https://github.com/matthieusieben)! - Build system rework, stop bundling dependencies.
+
+### Patch Changes
+
+- Updated dependencies [[`f689bd51a`](https://github.com/bluesky-social/atproto/commit/f689bd51a2f4e02d4eca40eb2568a1fcb95494e9)]:
+  - @atproto/common@0.4.0
+  - @atproto/crypto@0.4.0
+  - @atproto/repo@0.4.0
+
 ## 0.1.9
 
 ### Patch Changes