+++ packages/lex-cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atproto/lex-cli
 
+## 0.4.0
+
+### Minor Changes
+
+- [#2169](https://github.com/bluesky-social/atproto/pull/2169) [`f689bd51a`](https://github.com/bluesky-social/atproto/commit/f689bd51a2f4e02d4eca40eb2568a1fcb95494e9) Thanks [@matthieusieben](https://github.com/matthieusieben)! - Build system rework, stop bundling dependencies.
+
+### Patch Changes
+
+- Updated dependencies [[`f689bd51a`](https://github.com/bluesky-social/atproto/commit/f689bd51a2f4e02d4eca40eb2568a1fcb95494e9)]:
+  - @atproto/lexicon@0.4.0
+  - @atproto/syntax@0.3.0
+
 ## 0.3.2
 
 ### Patch Changes